pyproject.toml
[project]
name = "qa-answer-streamlit"
version = "0.2.0"
description = "PDF Q&A Chatbot with Hybrid RAG (BM25 + Semantic)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.4",
    "langchain>=1.2.9",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
]
